% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterYear.R
\name{filter_by_year}
\alias{filter_by_year}
\title{Filter Spending Data by Year Range (This is the function for for extra mark :))}
\usage{
filter_by_year(data, start_year, end_year)
}
\arguments{
\item{data}{Data frame containing spending data.}

\item{start_year}{The starting year for filtering.}

\item{end_year}{The ending year for filtering.}
}
\value{
A filtered data frame containing data between start_year and end_year.
}
\description{
Filter Spending Data by Year Range (This is the function for for extra mark :))
}
