% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_spending}
\alias{historical_spending}
\title{Valentine's Day Consumer Spending Data from 2010 to 2022}
\format{
\subsection{'historical_spending'}{

A data frame with 13 rows and 10 columns:
\describe{
\item{Year}{Year of recorded spending (2010-2022).}
\item{PercentCelebrating}{Percentage of people celebrating Valentine's Day in the given year.}
\item{PerPerson}{Average amount spent per person in USD.}
\item{Candy}{Average amount spending on candy in USD.}
\item{Flowers}{Average amount spending on flowers in USD.}
\item{Jewelry}{Average amount spending on jewelry in USD.}
\item{GreetingCards}{Average amount spending on greeting cards in USD.}
\item{EveningOut}{Average amount spending on evening outings in USD.}
\item{Clothing}{Average amount spending on clothing in USD.}
\item{GiftCards}{Average amount spending on gift cards in USD.}
}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2024/2024-02-13}
}
\usage{
historical_spending
}
\description{
A dataset containing historical consumer spending trends on Valentine's Day in the United States across multiple categories, from the years 2010 to 2022.
}
\keyword{datasets}
