% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_valentine_app}
\alias{launch_valentine_app}
\title{Launch the Valentine's Day Consumer Trends Shiny App}
\usage{
launch_valentine_app()
}
\description{
This function launches the Shiny app included in the valentineTool package.
}
