% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trends_custom}
\alias{plot_trends_custom}
\title{Plot Multiple Consumer Spending Trends with Custom Colors}
\usage{
plot_trends_custom(data, y_vars)
}
\arguments{
\item{data}{A data frame containing the historical spending data.}

\item{y_vars}{A vector of strings representing the columns to be plotted on the Y-axis.}
}
\value{
A ggplot object
}
\description{
This function plots the selected Y-axis variables over time, each with a different color.
}
